+++ lib/presentation/controllers/url_launcher.dart
@@ -5,7 +5,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../../../domain/entities/url_launcher.dart';
+import '../../domain/entities/url_launcher.dart';
 
 /// URL起動プロバイダー
 ///
