+++ lib/presentation/pages/repo/components/repo_detail_view.dart
@@ -8,11 +8,11 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../../domain/repositories/repo/entities/repo.dart';
 import '../../../../utils/logger.dart';
 import '../../../../utils/routing/router.dart';
-import '../../../components/cached_circle_avatar.dart';
-import '../../../components/error_view.dart';
-import '../../../components/hyperlink_text.dart';
-import '../../../components/icon_label.dart';
-import '../../../components/states/url_launcher.dart';
+import '../../../controllers/url_launcher.dart';
+import '../../../widgets/cached_circle_avatar.dart';
+import '../../../widgets/error_view.dart';
+import '../../../widgets/hyperlink_text.dart';
+import '../../../widgets/icon_label.dart';
 import 'readme_markdown.dart';
 import 'repo_language_label.dart';
 import 'selected_repo.dart';
