+++ lib/presentation/pages/repo/components/repo_full_name_text.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../domain/repositories/repo/entities/repo.dart';
-import '../../../components/async_value_handler.dart';
+import '../../../widgets/async_value_handler.dart';
 import 'selected_repo.dart';
 
 /// リポジトリ名を表示するText
