+++ lib/presentation/pages/repo/components/repo_list_view.dart
@@ -15,10 +15,10 @@ import '../../../../utils/assets/assets.gen.dart';
 import '../../../../utils/localizations/strings.g.dart';
 import '../../../../utils/logger.dart';
 import '../../../../utils/routing/router.dart';
-import '../../../components/cached_circle_avatar.dart';
-import '../../../components/error_view.dart';
-import '../../../components/icon_label.dart';
-import '../../../components/list_loader.dart';
+import '../../../widgets/cached_circle_avatar.dart';
+import '../../../widgets/error_view.dart';
+import '../../../widgets/icon_label.dart';
+import '../../../widgets/list_loader.dart';
 import 'repo_language_label.dart';
 import 'search_repos.dart';
