+++ lib/presentation/widgets/error_view.dart
