+++ test/presentation/components/search_app_bar_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/presentation/components/search_app_bar.dart';
+import 'package:github_search/presentation/widgets/search_app_bar.dart';
 
 import '../../test_utils/test_agent.dart';
